backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dash"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
sqlmodel = "^0.0.21"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
rich = "^13.7.1"
twilio = "^9.3.8"
pydantic = {version = "2.8.2", extras = ["email"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
